+++ section-parts/section-news.php
@@ -33,12 +33,28 @@ $desc = get_theme_mod( 'onepress_news_desc' );
 				<div class="col-sm-12">
 					<div class="blog-entry wow slideInUp">
 						<?php
-						$query = new WP_Query(
-							array(
-								'posts_per_page' => $number,
-								'suppress_filters' => 0,
-							)
-						);
+
+                        $cat = absint( get_theme_mod( 'onepress_news_cat' ) );
+                        $orderby = sanitize_text_field( get_theme_mod('onepress_news_orderby') );
+                        $order = sanitize_text_field( get_theme_mod('onepress_news_order') );
+
+                        $args = array(
+                            'posts_per_page' => $number,
+                            'suppress_filters' => 0,
+                        );
+                        if ( $cat > 0 ) {
+                            $args['category__in'] = array( $cat );
+                        }
+
+                        if ( $orderby && $orderby != 'default' ) {
+                            $args['orderby'] = $orderby;
+                        }
+
+                        if ( $order) {
+                            $args['order'] = $order;
+                        }
+
+						$query = new WP_Query( $args );
 						?>
 						<?php if ( $query->have_posts() ) : ?>